Abstract constraint validation
